CLI Templating, Frontend Upgrade started
-- Install Markdown Guide --
*Fixed some formatting issues, added two pretty important dependencies
PHP needs to have installed, and Adminer, an alternative to PHPMyAdmin.

-- Front End --
Largely modernized the structure and design, and brought most of the old
modules to the new CP strucutre (which includes a menu of all available
Control Panel pages, too!).
Added lots of relevant design files, like fonts.
Recently updated/fixed/redesigned modules are: media, galleries, mails,
plugins, CPMenu.

-- Settings --
Added a new type of settings - Meta settings. Those settings are meant
to be used by the settings API, to signify which settings are available
(or at least, available through the API).
Added to the API, the intaller, etc.

-- FileHandler --
Added verbose, and a few new params to the writing mode, such as whether
to override existing files, or try to create folders leading up to a
destination.

-- CLI --
*Small plugins fix
*Added a whole new module for frontend templating.

-- Util --
* Old siteNameCleanse became an IOFrame util function file,
modificationFunctions. The naming of functions slightly changed.
* Added templateFunctions, which (for now) contains a simple function to
support templating.

-- Templating --
Added a whole module to the cli tools, called "frontend", which focuses
on creating new pages (and all the relevant css/js) in the current
IOFrame page structure.