Relog with cookies
-- Auto Relog with Cookies --
Added a setting (in userSettings) that allows to relogging using
cookies.
Basically, if set, it will affect three things:
1. During UserHandler->logIn(), if the loging succeeds and userID is
set, cookies for relog will be set.
2. In coreInit, if you are not log in but relog cookies are set, will
try to log you in, and update the cookies according to the result.
3. UserHandler->logOut(), if you are not logging somebody else out, all
those cookies are erased.
After some testing this seems to be working perfectly.