1.1.1.0
---- Tokens
* Added tags to tokens. Can now search for all tokens containing said tags. Updated DB Table, Handler, API, and CP module.
* Fixed a small bug in TokenHandler

---- User Invites
User invites are now live. Invites allow a user to register, even when registration is regularly closed.
Invites can be sent via mail, or generated and later printed via a QR code.
Typically, QR codes are meant for general registration, possibly with more than one use, while mail invites are meant to be used once and are email restricted.
Invites are implemented above on top of tokens.
You can see all invites in the Users CP module, in "Tokens" mode (where they can also be created, and mailed if need be).
Updated UserHandler, API, and CP module.
Registraction component updated, and the default LoginRegister module (and page) updated as well.
Created a new default invitation mail template (although, as is the case with the others templates, it REALLY needs to be updated for proper systems).

---- Fixes
* Small searchList component fix.
* Small toggle2faApp component fix.
* Small tokensEditor component fix.
* Update module no longer in test mode by default.
* Updated some tests.